bkmks/bkmk_class/requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
appdirs==1.4.4
apsw==3.35.4.post1
arrow==1.1.0
asgiref==3.3.4
astroid==2.5.6
async-timeout==3.0.1
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
bcj-cffi==0.5.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
black==21.4b2
blessings==1.7
blinker==1.4
BookCut==1.3.4
bottle==0.12.19
Brlapi==0.8.2
btrfsutil==5.13
CacheControl==0.12.6
cchardet==2.1.7
cffi==1.14.6
chardet==4.0.0
cheroot==8.5.2
CherryPy==18.6.0
click==7.1.2
colorama==0.4.4
Colr==0.9.1
commonmark==0.9.1
constantly==15.1.0
contextlib2==0.6.0.post1
cookiecutter==1.7.2
coverage==5.5
cryptography==3.4.7
css-parser==1.0.6
cssselect==1.1.0
curtsies==0.3.5
cwcwidth==0.1.4
cycler==0.10.0
dateutils==0.6.12
decorator==5.0.7
distlib==0.3.2
distro==1.5.0
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
dnspython==2.1.0
docutils==0.17.1
epy-reader==2021.4.7
fabric==2.6.0
feedgenerator==1.9.1
feedparser==6.0.2
filelock==3.0.12
fire==0.4.0
flake8==3.9.1
Flask==1.1.2
greenlet==1.0.0
gunicorn==20.1.0
html2text==2020.1.16
html5-parser==0.4.9
html5lib==1.1
hupper==1.10.2
hyperlink==21.0.0
idna==3.2
ifaddr==0.1.7
incremental==21.3.0
iniconfig==1.1.1
invoke==1.5.0
ipython==7.23.0
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==1.1.0
jaraco.classes==3.2.1
jaraco.collections==3.3.0
jaraco.functools==3.3.0
jaraco.text==3.5.0
jedi==0.18.0
Jinja2==2.11.3
jinja2-time==0.2.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
libfdt==1.6.1
libgenesis==0.1.7
libvirt-python==7.3.0
loguru==0.5.3
louis==3.18.0
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
md2html==1.3.0
mechanize==0.4.5
mobi==0.3.1
more-itertools==8.7.0
msgpack==1.0.2
multidict==5.1.0
multivolumefile==0.2.3
mypy-extensions==0.4.3
mysql-connector==2.2.9
netifaces==0.11.0
nftables==0.1
nord==0.2.0
numpy==1.20.3
ordered-set==4.0.2
packaging==20.9
pandas==1.2.4
paramiko==2.7.2
parso==0.8.2
PasteDeploy==2.1.1
pathlib2==2.3.5
pathspec==0.8.1
pdfminer.six==20201018
pdftotext==2.1.6
pdfx==1.4.1
pelican==4.6.0
pep517==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plaster==1.0
plaster-pastedeploy==0.7
pluggy==0.13.1
ply==3.11
portend==2.7.1
poyo==0.5.0
progress==1.5
prompt-toolkit==3.0.18
psutil==5.8.0
psycopg2==2.9.1
ptpython==3.0.17
ptyprocess==0.7.0
pwquality==1.4.4
py==1.10.0
py7zr==0.15.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
pychm==0.8.6
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pydub==0.25.1
pyfiglet==0.8.post1
pyflakes==2.3.1
Pygments==2.9.0
PyGObject==3.40.1
pylint==2.8.2
pymongo==3.12.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyppmd==0.12.1
PyQt5==5.15.4
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.4
pyramid==2.0
pyramid-debugtoolbar==4.9
pyramid-jinja2==2.8
pyramid-mako==1.1.0
pytest==6.2.3
pytest-cov==2.11.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-slugify==5.0.0
pytz==2021.1
pyxdg==0.27
pyzstd==0.14.4
questionary==1.9.0
queuelib==1.6.1
ranger-fm==1.9.3
regex==2021.7.6
repoze.lru==0.7
requests==2.26.0
requests-file==1.5.1
resolvelib==0.5.5
retrying==1.3.3
rich==10.1.0
Scrapy==1.0.3
service-identity==21.1.0
sgmllib3k==1.0.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
structlog==21.1.0
sultan==0.9.1
tabulate==0.8.9
team==1.0
tempora==4.0.2
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.4
tldextract==3.1.0
toml==0.10.2
tomli==1.2.0
tqdm==4.61.0
traitlets==5.0.5
translationstring==1.4
Twisted==21.7.0
typing-extensions==3.7.4.3
Unidecode==1.2.0
unrardll==0.1.4
urllib3==1.26.6
urwid==2.1.2
venusian==3.0.0
virtualenv==20.4.4
w3lib==1.22.0
waitress==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
WebTest==2.0.35
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
yarl==1.6.3
zc.lockfile==2.0
zeroconf==0.29.0
zope.deprecation==4.4.0
zope.interface==5.4.0
